Split visualizer code into own file to comply with license
